backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
email-validator
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
google-generativeai>=0.8.0,<1.0.0
Pillow==10.1.0
python-magic==0.4.27
